[feat] Improved memory management processing windows as they are read
Previously to this commit, the software stored imu data in a big matrix, which constantly increased until the status of the software changed to "Processing". This lead to unexpected behaviours since the memory usage easily reached the full capacity of the arduino.

It is necessary to proccess the data as it is read so that we don't need to store a huge matrix in memory.

This new approach, can lead to loose some records because while the software is processing a small set of data, it cannot read new data. However, after some tests we can conclude that the data loss is so small that can be accepted. For example, with a window_size of 70, we loose about 4 records between one window and another.